+++ lib/surrogate.rb
@@ -10,8 +10,8 @@ require 'surrogate/invocation'
 class Surrogate
   UnpreparedMethodError = Class.new StandardError
 
-  def self.endow(klass, &block)
-    Endower.endow klass, &block
+  def self.endow(klass, options={},  &block)
+    Endower.endow klass, options, &block
     klass
   end
 end